+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/client/AbstractHttpClient.java
@@ -1,7 +1,10 @@
 package io.github.clescot.kafka.connect.http.client;
 
+import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
+import com.google.common.collect.Maps;
 import dev.failsafe.RateLimiter;
+import dev.failsafe.RateLimiterConfig;
 import io.github.clescot.kafka.connect.http.core.HttpExchange;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,11 +21,14 @@ import java.io.InputStream;
 import java.nio.file.Path;
 import java.security.*;
 import java.security.cert.CertificateException;
+import java.time.Duration;
+import java.time.temporal.ChronoUnit;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
 import static io.github.clescot.kafka.connect.http.client.Configuration.CONFIGURATION_ID;
+import static io.github.clescot.kafka.connect.http.client.Configuration.STATIC_SCOPE;
 import static io.github.clescot.kafka.connect.http.sink.HttpSinkConfigDefinition.*;
 
 public abstract class AbstractHttpClient<Req, Res> implements HttpClient<Req, Res> {
@@ -32,10 +38,15 @@ public abstract class AbstractHttpClient<Req, Res> implements HttpClient<Req, Re
     private Optional<RateLimiter<HttpExchange>> rateLimiter = Optional.empty();
     protected TrustManagerFactory trustManagerFactory;
     protected  String configurationId;
+
+    //rate limiter
+    private final static Map<String, RateLimiter<HttpExchange>> sharedRateLimiters = Maps.newHashMap();
+
     protected AbstractHttpClient(Map<String, Object> config) {
         this.config = config;
         configurationId = (String) config.get(CONFIGURATION_ID);
         Preconditions.checkNotNull(configurationId,"configuration must have an id");
+        setRateLimiter(buildRateLimiter(config));
     }
 
     protected Optional<TrustManagerFactory> buildTrustManagerFactory() {
@@ -168,5 +179,45 @@ public abstract class AbstractHttpClient<Req, Res> implements HttpClient<Req, Re
         }
         result.append("}");
         return result.toString();
+    }
+
+    private RateLimiter<HttpExchange> buildRateLimiter(Map<String, Object> configMap) {
+        String configurationId = (String) configMap.get(CONFIGURATION_ID);
+        RateLimiter<HttpExchange> rateLimiter = null;
+        if (configMap.containsKey(RATE_LIMITER_MAX_EXECUTIONS)) {
+            long maxExecutions = Long.parseLong((String) configMap.get(RATE_LIMITER_MAX_EXECUTIONS));
+            LOGGER.trace("configuration '{}' : maxExecutions :{}",configurationId,maxExecutions);
+            long periodInMs = Long.parseLong(Optional.ofNullable((String) configMap.get(RATE_LIMITER_PERIOD_IN_MS)).orElse(1000 + ""));
+            LOGGER.trace("configuration '{}' : periodInMs :{}",configurationId,periodInMs);
+            if (configMap.containsKey(RATE_LIMITER_SCOPE) && STATIC_SCOPE.equalsIgnoreCase((String) configMap.get(RATE_LIMITER_SCOPE))) {
+                LOGGER.trace("configuration '{}' : rateLimiter scope is 'static'",configurationId);
+                Optional<RateLimiter<HttpExchange>> sharedRateLimiter = Optional.ofNullable(sharedRateLimiters.get(configurationId));
+                if (sharedRateLimiter.isPresent()) {
+                    rateLimiter = sharedRateLimiter.get();
+                    LOGGER.trace("configuration '{}' : rate limiter is already shared",configurationId);
+                } else {
+                    rateLimiter = RateLimiter.<HttpExchange>smoothBuilder(maxExecutions, Duration.of(periodInMs, ChronoUnit.MILLIS)).build();
+                    registerRateLimiter(configurationId, rateLimiter);
+                }
+            } else {
+                rateLimiter = RateLimiter.<HttpExchange>smoothBuilder(maxExecutions, Duration.of(periodInMs, ChronoUnit.MILLIS)).build();
+            }
+            RateLimiterConfig<HttpExchange> config = rateLimiter.getConfig();
+            LOGGER.trace("configuration '{}' rate limiter configured : 'maxRate:{},maxPermits{},period:{},maxWaitTime:{}'",configurationId,config.getMaxRate(),config.getMaxPermits(),config.getPeriod(),config.getMaxWaitTime());
+        }else{
+            if(LOGGER.isTraceEnabled()) {
+                LOGGER.trace("configuration '{}' : rate limiter is not configured", configurationId);
+                LOGGER.trace(Joiner.on(",\n").withKeyValueSeparator("=").join(configMap.entrySet()));
+            }
+        }
+        return rateLimiter;
+    }
+
+
+    public static void registerRateLimiter(String configurationId, RateLimiter<HttpExchange> rateLimiter) {
+        Preconditions.checkNotNull(configurationId, "we cannot register a rateLimiter for a 'null' configurationId");
+        Preconditions.checkNotNull(rateLimiter, "we cannot register a 'null' rate limiter for the configurationId " + configurationId);
+        LOGGER.info("registration of a shared rateLimiter for the configurationId '{}'", configurationId);
+        sharedRateLimiters.put(configurationId, rateLimiter);
     }
 }
